Rework of code and enough for 1.0.0 beta
The whole code base has been reworked but still has the same core
methods. The changed core for the database connections is now a generic
interface, these are called model drivers, the implementation of these
upholds the connection and takes the generic data into database specific
commands. The core active record is now called a model and has become a
interface instead of a abstract class. Metadata about the
columns/properties are now handled directly on instantiation of the
value wrapper.

An addition that came out of the rework is what I call compotites. A
composite are like views in a SQL database. You can get
columns/properties from other models and for now only count the amount
of items in a reference list. (see the social post media example)

I have built this new code base on the basis that it should be
extensible by the user. There are some stuff left for that in the way I
handle model properties. That is why it is in beta as I am confident in
the current API but have no real big changes in mind.